requirements.txt
Werkzeug==0.14.1
Flask==1.0.2
Pillow==5.3.0
pymongo==3.7.2